requirements.txt
# Parsec Zero — Automated Game Studio
# Python 3.11+ required

# Agent framework
crewai>=0.28.0
crewai-tools>=0.1.0

# Anthropic LLM
anthropic>=0.25.0
langchain-anthropic>=0.1.0
langchain-core>=0.1.0

# Vector memory
chromadb>=0.4.0

# Environment & config
python-dotenv>=1.0.0

# Git automation
gitpython>=3.1.0

# HTTP requests (Stable Diffusion, Suno, ElevenLabs)
requests>=2.31.0
httpx>=0.27.0

# Asset handling
pillow>=10.0.0

# Scheduling and rate limiting
apscheduler>=3.10.0
tenacity>=8.2.0

# Logging
structlog>=24.0.0

# CLI
click>=8.1.0
rich>=13.0.0
